--- components/dnd/Droppable.tsx
@@ -1,19 +1,0 @@
-import { useDroppable } from '@dnd-kit/core';
-
-export default function Droppable(props: {
-  id: string;
-  children: React.ReactNode;
-}) {
-  const { isOver, setNodeRef } = useDroppable({
-    id: props.id,
-  });
-  const style = {
-    color: isOver ? 'green' : undefined,
-  };
-
-  return (
-    <div ref={setNodeRef} style={style}>
-      {props.children}
-    </div>
-  );
-}
