+++ components/formCreator/FormCreator.tsx
@@ -8,7 +8,6 @@ import {
   IconSettings,
   IconX,
 } from '@tabler/icons-react';
-import classes from '@/styles/Input.module.css';
 import { v4 as uuidv4 } from 'uuid';
 import {
   FormSection,
