+++ styles/Input.module.css
@@ -1,18 +1,3 @@
-.input {
-  width: 100%;
-  border-width: 0 0 2px 0;
-  border-style: solid;
-  border-radius: 0;
-  padding-left: 0;
-  padding-right: 0;
-  background-color: transparent;
-
-  &:focus {
-    outline: none;
-    border-color: var(--mantine-color-blue-3);
-  }
-}
-
 .title {
   margin-top: 1rem;
   width: 100%;
